Task313_vers_5

+++ src/main/java/ru/kata/spring/boot_security/demo/controllers/AdminController.java
@@ -15,6 +15,7 @@ import org.springframework.web.bind.annotation.RequestParam;
 import ru.kata.spring.boot_security.demo.model.User;
 import ru.kata.spring.boot_security.demo.service.AdditionalService;
 import ru.kata.spring.boot_security.demo.service.UserService;
+import ru.kata.spring.boot_security.demo.service.UserServiceImpl;
 import ru.kata.spring.boot_security.demo.validators.UserValidator;
 
 import javax.validation.Valid;
@@ -27,10 +28,7 @@ public class AdminController {
     private final AdditionalService additionalService;
     private final UserValidator userValidator;
 
-    @Autowired
-    public AdminController(UserService userService,
-                           AdditionalService additionalService,
-                           UserValidator userValidator) {
+    public AdminController(UserService userService, AdditionalService additionalService, UserValidator userValidator) {
         this.userService = userService;
         this.additionalService = additionalService;
         this.userValidator = userValidator;
@@ -40,11 +38,33 @@ public class AdminController {
     public String showAllUsers(Model model, Principal principal) {
         model.addAttribute("newUser", new User());
         additionalService.createModelForView(model, principal);
+        model.addAttribute("users", userService.getAllUsers());
         model.addAttribute("activeTab", "usersTable");
         return "adminPage";
     }
 
-    @PostMapping("/admin")
+    @GetMapping("/edit/{id}")
+    public String showEditUserForm(@PathVariable("id") int id, Model model) {
+        User user = userService.showUserById(id);
+        model.addAttribute("userIter", user);
+        return "editUser";
+    }
+
+    @PostMapping("/admin/update") // Изменено на /admin/update
+    public String updateUser(@ModelAttribute("userIter") @Valid User user,
+                             BindingResult bindingResult,
+                             Model model, Principal principal) {
+        model.addAttribute("authUser", userService.findByUsername(principal.getName()));
+
+        if (bindingResult.hasErrors()) {
+            return "adminPage"; // Вернуть на страницу админки в случае ошибок
+        }
+
+        userService.updateUser(user); // Обновляем пользователя
+        return "redirect:/users/admin"; // Перенаправление на страницу админки после успешного обновления
+    }
+
+    @PostMapping("/admin/add") // Изменено на /admin/add
     public String addUser(@ModelAttribute("newUser") @Valid User user, BindingResult bindingResult,
                           Principal principal, Model model) {
         userValidator.validate(user, bindingResult);
@@ -57,36 +77,27 @@ public class AdminController {
         return "redirect:/users/admin";
     }
 
-    @PatchMapping("/admin")
-    public String updateUser(@ModelAttribute("userIter") @Valid User user,
-                             BindingResult bindingResult,
-                             Model model, Principal principal) {
-        model.addAttribute("authUser", userService.findByUsername(principal.getName()));
-        if (bindingResult.hasErrors()) {
-            return "adminPage";
+    @PostMapping("/admin/delete") // Изменено на /admin/delete
+    public String deleteUser(@RequestParam("id") int id) {
+        if (id > 0) {
+            userService.deleteById(id);
         }
-        userService.updateUser(user);
         return "redirect:/users/admin";
     }
 
-    @DeleteMapping("/admin")
-    public String deleteUser(Model model, @RequestParam("id") int id) {
-        model.addAttribute("user", userService.showUserById(id));
-        userService.deleteById(id);
-        return "redirect:/users/admin";
-    }
 
-    @GetMapping("/edit/{id}")
-    public String showEditUserForm(@PathVariable("id") int id, Model model) {
-        User user = userService.showUserById(id);
-        model.addAttribute("userIter", user);
-        return "editUser";
-    }
+  /*
+
+
+
+
+
+
 
     @GetMapping("/delete/{id}")
     public String confirmDeleteUser(@PathVariable("id") int id, Model model) {
         User user = userService.showUserById(id);
         model.addAttribute("user", user);
         return "confirmDelete";
-    }
+    }*/
 }
